TFNS M3: optimizer-safe update (project the applied Adam delta) + isolation fix
Spec §13 hard invariant + §23.5/§23.6.
- protect/optimizer.py: exact §13 sequence — project raw grad, ask Adam for a
  candidate delta+state, PROJECT THE APPLIED DELTA, project Adam first moments
  (optax.tree_utils with namedtuple fallback), norm bound, backtracking
  accept/reject with full params+opt-state restore on reject; telemetry info.
  constraint_fn/accept_fn hooks for M5. No decoupled weight decay.
- 5/5 safety tests: GRU sequence invariance under projected update (incl resets;
  strengthened control with order-of-magnitude separation), Adam raw-grad
  projection is insufficient while applied-delta projection restores invariance,
  first-moment projection, reject restores optimizer state, accept + norm bound.
- Test isolation: removed module-level jax_enable_x64 (it leaked globally and
  broke other modules); all tfns tests run float32 with 1e-4 tolerances; added
  conftest x64-off guard; made unroll scan carry dtype-consistent.

+++ tfns/model/agent.py
@@ -224,7 +224,7 @@ class RecurrentAgent(nn.Module):
                 ema_encoder_params=ema_encoder_params,
                 collect_presyn=collect_presyn,
             )
-            return out.h_next, out
+            return out.h_next.astype(h_prev.dtype), out
 
         h_final, outputs = jax.lax.scan(step, h0, (obs_seq, act_seq, rew_seq, reset_seq))
         return outputs, h_final
